Beginning to simplify Write-StatusBar
This was using Write-Color, which I don't remember writing. Internally it used Write-Host, giving up lots of control--this meant mocking Write-Host to capture data. Beginning to use Format-Output so it just returns strings, which is much nicer to assert on. It also means I can glue the whole string together and output the whole thing (better performance) instead of sending output to the console multiple times in a row.