Now generates a test to show that there are only N exposed functions
This ensures that as the public API surface changes (especially if things are removed), a test is triggered. Removing commands is broken backward compatibilty, and this'll catch it. This also will error if new functions are added. In this case, the test should be regenerated, since addition is OK, and that would be the new stable normal.