Fixing Get-FileDetail to return one single hashtable
Each reader contributes to a single one at the end, instead of directly sending a bunch of hashtables through. It works out at the end of the pipeline for display, but it makes the tests messier. We'll try this for a while and see how it goes. Duplicate keys are now an issue though.