+Menu
This is a high-level "UI control" for the console. Controls like this have their own Region to define their visual appearance. So far it has its own interaction logic (key input stuff). There's likely better ways to do this in the future in how responsibilities are separated, but for now this works.