+Color abstraction
It'd make sense to use ConsoleColor, except by making our own we can add a "transparent" option to indicate that the region's cell attribute should NOT be altered at all. Without this, it'd be much harder to read the attribute for a character we don't want to modify, only to blanket write all attributes.