+ColorCell struct
This encapsulates the data for one "cell" -- a character, a foreground color, and a background color. This way we can store an array of these items, rather than storing discrete arrays of differing data (chars vs. attributes). This helps with locality of reference, but also lets us deal in complete objects in one spot.