+Test project for "Core" assembly
This will be the new name for "native." This is because "native" is too specific to the implementation. It's not about interop or native code; it's about the lowest level of the library. It's also because the "native" project has more than native stuff, and is accumulating an API surface that exposes a higher level of ideas than just interop.