Input elements can now enter an input loop and request redraws
The Use method enters an input frame that captures input and then passes the info to its abstract input handler, overridden by subclasses, with the key info that was received. This allows inheritors to get plugged into an input loop for free.

The Invalidate method calls the OnPaint to give subclasses a chance to add their custom painting using the region context. It then raises the RequestDraw event to signal to the framework that it should be redrawn to the screen.