No longer crashes sometimes on long menu items
It was accommodating the smallest, not the largest--no idea what I was thinking here. It also hard-codes the margin to accommodate the indentation for selection indicator. This should have a configurable margin, not hard code it. It also should be smart enough not write past the end of the array.