Can now request the render manager to repaint
This now causes the next cycle to print the changes. This allows for an on-demand painting, but it may make more sense to have this thing chug along at a predictable frame rate and just have it paint on its own every time. Or maybe do it lazily if there are changes? We'll see--this is the first step.