Beginning to change how animations work, separating rasterization
Animations should operate on TextBuffer data and not concern themselves with formatting VT sequences (like the cursor animation had been doing to fill in the background color). We want these data classes to animate data, and then the final step in rendering is that the TextBuffer is "flattened" to VT sequences and printed, and this step happens last. It also isolates this flattening responsibility to rendering, and not animations or other data crunchers.