+ITextRasterizer
This is the parent thing that knows how to turn a TextBuffer into a string that can be printed wherever it needs to go--probably the console. This abstraction gives us space to have different kinds, though the default one we'll roll with is the upcoming rasterizer that turns colors into VT sequences.