Standing up the tree and the text node
Right now the FormatTree does not scan the TextBuffer; the current algorithm takes the whole string and puts it into one TextNode, who evaluates to its own text. This is enough to keep the single test passing (for text only) and does not accommodate color changes. It's enough to stand up the objects I'll need.