+Private FormatState class
We'll use this to track the current state and detect when it changes. This takes the place of a pile of loose local variables. There may be a way to extract this data more easily from a Cell rather than all this manual work, but this is an OK first step for now.