+FormatBuilder and test class
A stateful way to alter the formatting for output. This is pretty similar to the FormatTree and node stuff I built earlier, and this may be a simpler way to do it. This is also a more approachable API that could be usable in other parts of the code. Not sure whether this totally replaces that other stuff yet, or maybe works with it.