Reworking FormatBuilder to use the nodes
This ditches the dumb flag before it got too far. The Foreground node had to be updated to accommodate ConsoleColors as well as RGB. This seems like an OK place to put it, although I'd love an enum class or discriminated union to have one "color" type, whether it's a console color, RGB, or even indexed.